+++ src/commands/league/league-signup.ts
@@ -2,12 +2,14 @@ import { MemberCommand } from "../command";
 import { CustomInteraction } from "../interaction";
 import { isGuildMember } from "../../utility/utility";
 import { OverstatService } from "../../services/overstat";
-import { Snowflake, User } from "discord.js";
-import { GoogleAuth, OAuth2Client } from "googleapis-common";
-import { AnyAuthClient } from "google-auth-library";
-import { auth, sheets } from "@googleapis/sheets";
-import { SheetHelper } from "../../utility/sheet-helper";
-import { DB } from "../../db/db";
+import { User } from "discord.js";
+import {
+  LeagueSignupService,
+  SheetsPlayer,
+  PlayerRank,
+  Platform,
+  VesaDivision,
+} from "../../services/league-signup";
 
 export class LeagueSignupCommand extends MemberCommand {
   inputNames = {
@@ -43,7 +45,7 @@ export class LeagueSignupCommand extends MemberCommand {
 
   constructor(
     private overstatService: OverstatService,
-    private db: DB,
+    private leagueSignupService: LeagueSignupService,
   ) {
     super("league-signup", "Signup for the league");
     this.addStringInput(this.inputNames.teamName, "Team name", {
@@ -219,7 +221,7 @@ export class LeagueSignupCommand extends MemberCommand {
     await interaction.deferReply();
 
     try {
-      const signupNumber = await this.postSpreadSheetValue(
+      const signupNumber = await this.leagueSignupService.signup(
         teamName,
         teamNoDays ?? "Open schedule",
         compExperience,
@@ -324,124 +326,4 @@ export class LeagueSignupCommand extends MemberCommand {
     }
     return linkToReturn;
   }
-
-  async postSpreadSheetValue(
-    teamName: string,
-    teamNoDays: string,
-    teamCompKnowledge: string,
-    player1: SheetsPlayer,
-    player2: SheetsPlayer,
-    player3: SheetsPlayer,
-    additionalComments: string,
-  ): Promise<number | null> {
-    const authClient = await this.getAuthClient();
-
-    const returningPlayersCount = [player1, player2, player3].reduce(
-      (count, player) => {
-        if (player.previous_season_vesa_division !== VesaDivision.None) {
-          return count + 1;
-        } else {
-          return count;
-        }
-      },
-      0,
-    );
-    const values = [
-      [
-        new Date().toISOString(),
-        teamName,
-        teamNoDays,
-        teamCompKnowledge,
-        `${returningPlayersCount} returning players`,
-        ...this.convertSheetsPlayer(player1),
-        ...this.convertSheetsPlayer(player2),
-        ...this.convertSheetsPlayer(player3),
-        additionalComments,
-      ],
-    ];
-
-    const sheetsClient = sheets({ version: "v4" });
-    const activeSeason = await this.db.getActiveLeagueSeason();
-    if (!activeSeason) {
-      throw new Error("No season found with active signups.");
-    }
-    const request = SheetHelper.BUILD_REQUEST(
-      values,
-      authClient as OAuth2Client,
-      {
-        id: activeSeason.googleSheetId,
-        range: `${activeSeason.googleSheetName}!${activeSeason.googleSheetRangeStart}`,
-      },
-    );
-
-    const response = await sheetsClient.spreadsheets.values.append(request);
-    console.log(response.data);
-    const rowNumber = SheetHelper.GET_ROW_NUMBER_FROM_UPDATE_RESPONSE(
-      response.data.updates,
-    );
-    return rowNumber ? rowNumber - SheetHelper.STARTING_CELL_OFFSET : null;
-  }
-
-  private convertSheetsPlayer(player: SheetsPlayer): (string | number)[] {
-    return [
-      player.name,
-      player.discordId,
-      player.overstatLink ?? "No overstat",
-      VesaDivision[player.previous_season_vesa_division],
-      PlayerRank[player.rank],
-      Platform[player.platform],
-      player.elo ?? "No elo on record",
-    ];
-  }
-
-  getAuthClient(): Promise<AnyAuthClient> {
-    const googleAuth = new auth.GoogleAuth({
-      keyFile: "service-account-key.json",
-      scopes: ["https://www.googleapis.com/auth/spreadsheets"],
-    }) as GoogleAuth;
-
-    return googleAuth.getClient();
-  }
-}
-
-enum PlayerRank {
-  Bronze,
-  Silver,
-  Gold,
-  Plat,
-  LowDiamond,
-  HighDiamond,
-  Masters,
-  Pred,
-}
-
-enum VesaDivision {
-  None,
-  Division1,
-  Division2,
-  Division3,
-  Division4,
-  Division5,
-  Division6,
-  Division7,
-  // Division8,
-  // Division9,
-  // Division10,
-}
-
-enum Platform {
-  pc,
-  playstation,
-  xbox,
-  switch,
-}
-
-interface SheetsPlayer {
-  name: string;
-  discordId: Snowflake;
-  elo: number | undefined;
-  rank: PlayerRank;
-  previous_season_vesa_division: VesaDivision;
-  platform: Platform;
-  overstatLink: string | undefined;
 }
