+++ src/services/index.ts
@@ -10,6 +10,7 @@ import { DiscordService } from "./discord";
 import { client } from "../Client";
 import { BanService } from "./ban";
 import { HuggingFaceService } from "./hugging-face";
+import { LeagueSignupService } from "./league-signup";
 
 // This file creates all the singleton services
 export const huggingFaceService = new HuggingFaceService();
@@ -43,3 +44,4 @@ export const rosterService = new RosterService(
   scrimService,
   signupsService,
 );
+export const leagueSignupService = new LeagueSignupService(nhostDb);
